refactored ServerResponseHandler in zhttp.service
